+++ src/renderer/js/views/pipeline/designer/strategies/DrawStrategy.js
@@ -50,13 +50,7 @@ export class DrawStrategy extends InteractionStrategy {
         // Update cursor
         this.controller.canvas.style.cursor = this.getCursor();
 
-        // Update hover state
-        const overNode = this.controller.findNodeAt(worldPos);
-        const newHoverId = overNode ? overNode.id : null;
-        if (this.controller.hoveredNodeId !== newHoverId) {
-            this.controller.hoveredNodeId = newHoverId;
-            this.controller.onUpdate?.();
-        }
+
     }
 
     /**